refactor(scoring): move BoardScore and Multiplier to Models.Scores namespace
