+++ lib/tests/DartsScorer.Tests/Scoring/LegTests.cs
@@ -1,5 +1,5 @@
 using DartsScorer.Main.Scoring.x01;
-using DartsScorer.Main.Scoring.Common;
+using DartsScorer.Main.Models.Scores;
 using DartsScorer.Main.Scoring;
 
 namespace DartsScorer.Tests.Scoring;
